Refactor New Flow
