+++ app/redux/map-view/mapview-my-visibility-saga.js
@@ -11,6 +11,7 @@ export default function* mapViewMyVisibilitySaga() {
     }
 
     yield takeEvery([
+      'MICRO_DATE_IM_READY',
       'MICRO_DATE_REMOVE',
       'MICRO_DATE_STOPPED_BY_TARGET',
       'MICRO_DATE_STOPPED_BY_ME',
